Added additional clauses to nix removal script

+++ shared/systems/optional/scripts/remove-nix.nix
@@ -6,8 +6,21 @@ let
      rm -rf $HOME/.nix-*
      rm -rf $HOME/.cache/nix
      rm -rf $HOME/.config/nixpkgs
-     sudo mv -f /etc/bashrc.backup-before-nix /etc/bashrc
+
+     if [ -e '/etc/bashrc.backup-before-nix' ]; then
+       sudo mv -f /etc/bashrc.backup-before-nix /etc/bashrc
+     fi
+
+     if [ -e '/etc/profile.d/nix.sh.backup-before-nix' ]; then
+       sudo mv -f /etc/profile.d/nix.sh.backup-before-nix /etc/profile.d/nix.sh
+     fi
+
+     if [ -e ' /etc/zshrc.backup-before-nix' ]; then
+     sudo mv -f /etc/zshrc.backup-before-nix /etc/zshrc
+     fi
+
      sudo rm -rf /nix
+
   '';
 in {
   environment.systemPackages = [ remove-nix ];
